+++ client/src/pages/Products/Products.js
@@ -5,38 +5,30 @@ import { Context } from "../../Contexts/Context";
 import LatestProduct from "../../components/LatestProduct/LatestProduct";
 import LoaderSpinner from "../../components/LoaderSpinner/LoaderSpinner";
 import Pagination from "../../components/Pagination/Pagination";
-import { useNavigate } from "react-router-dom";
 import ScrollToTop from "../../components/ScrollToTop/ScrollToTop";
+import usePaginationHook from "../../hooks/usePaginationHook";
+import useFetchData from "../../hooks/useFetchData";
 
 function Products() {
-  const { products, loading } = useContext(Context);
-
-  const navigate = useNavigate();
-
-  const [pageNumber, setPageNumber] = useState(0);
-
-  const PRODUCTS_PER_PAGE = 12;
-  const pagesVisited = pageNumber * PRODUCTS_PER_PAGE;
-  const pageCount = Math.ceil(products.length / PRODUCTS_PER_PAGE);
-
-  const handleChangePage = ({ selected }) => {
-    setPageNumber(selected);
-
-    navigate({
-      pathname: "/products",
-      search: selected !== 0 ? `?page=${selected + 1}` : null,
-    });
-  };
+  const apiProductEndpoint = `${process.env.REACT_APP_API_URI}/api/products`;
+  const { isLoaded, data } = useFetchData(apiProductEndpoint);
+  const {
+    pageNumber,
+    pagesVisited,
+    pageCount,
+    itemsPerPage,
+    handleChangePage,
+  } = usePaginationHook(0, data, 12, "/products");
 
   return (
     <div className={styles.Products}>
-      {loading ? (
-        <LoaderSpinner loading={loading} />
+      {isLoaded ? (
+        <LoaderSpinner loading={isLoaded} />
       ) : (
         <div className={styles.gridTemplate}>
           <div className={styles.itemsSection}>
-            {products
-              .slice(pagesVisited, pagesVisited + PRODUCTS_PER_PAGE)
+            {data
+              .slice(pagesVisited, pagesVisited + itemsPerPage)
               .map((item, index) => (
                 <LatestProduct
                   key={index}
